requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai
langgraph
google-generativeai
langchain-google-genai>=1.0.0
python-dotenv
sentence-transformers
numpy
faiss-cpu
matplotlib 
pandas
streamlit
